feat: update 1: modify some logics to integrate sdf map 2: remove lazythetastar planners 3: edit CMakeLists.txt for ROS and make it clean

+++ include/Planners/ThetaStar.hpp
@@ -80,7 +80,7 @@ namespace Planners
          * This operation of erase and re-insert is performed in order to update the position
          * of the node in the container. 
          */
-        virtual void exploreNeighbours(Node* _current, const Eigen::Vector3i &_target,node_by_position &_index_by_pos) override;
+        virtual void exploreNeighbours(Node* _current, const Eigen::Vector3d &_target,node_by_position &_index_by_pos) override;
 
         utils::CoordinateListPtr checked_nodes, checked_nodes_current;  /*!< TODO Comment */
